requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
